requirements.txt
pdfminer.six
PyMuPDF
pytesseract
Pillow
langdetect 
